+++ lib/Contracts.js
@@ -2,8 +2,8 @@
 
 const InstanceComposer = require('../instance_composer');
 
-require('../lib/contract_interacts/TokenHolder.js');
-require('../lib/contract_interacts/TokenRules.js');
+require('./contractInteracts/TokenHolder.js');
+require('./contractInteracts/TokenRules.js');
 
 const Contracts = function(config, ic) {
   const oThis = this;
