+++ lib/contractInteracts/TokenHolder.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const InstanceComposer = require('../../instance_composer');
-const generator = require('../../lib/contract_interacts/generator');
+const generator = require('./generator');
 
 const path = require('path'),
   fs = require('fs');
