+++ lib/contractInteracts/TokenRules.js
@@ -4,7 +4,7 @@ const path = require('path'),
   fs = require('fs');
 
 const InstanceComposer = require('../../instance_composer');
-const generator = require('../../lib/contract_interacts/generator');
+const generator = require('./generator');
 
 function parseFile(filePath, options) {
   filePath = path.join(__dirname, '/' + filePath);
