+++ lib/contractInteracts/generator.js
